consensus: allow one revault output during unvault
During the unvault process, allow some amount of the vault to be
"revaulted" into an output with the exact same scriptPubKey as the vault
that we're unvaulting from.

This is desirable because it allows the owner of the vault to manage the
remaining balance of the vault independently of the timelock governing
the initial unvault, which may be lengthy.

Idea from AJ Towns.