vaults: change recovery-spk-hash to recovery-params
Idea from Anthony Towns.

Instead of gating sweep-to-recovery by matching the output scriptPubKey
against a pre-specified hash, also include a scriptPubKey which must be
satisfied. I.e. instead of <recovery-spk-hash> being

  [32 bytes of recovery target scriptPubKey hash]

we use <recovery-params> instead:

  [32 bytes of recovery target scriptPubKey hash][recovery scriptPubKey]

To emulate the old behavior, the [recovery scriptPubKey] can be given as
OP_TRUE.

This gives users the option of avoiding sweep-to-recovery replay attacks
at the cost of having to maintain a key to authorize recovery sweeps.

Co-authored-by: Anthony Towns <aj@erisian.com.au>